+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog  
 All notable changes to this project will be documented in this file.
 
+## [0.21.1]
+
+### Fixed
+- Fixed setAppUserIdBase64
+
 ## [0.21.0]
 
 ### Added